# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medical Document Management System API",
    "description": "API documentation for the Medical Document Management System",
    "version": "1.0.0",
    "contact": {
      "email": "support@medicaldocs.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get system statistics (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_stats": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "integer"
                        },
                        "patients": {
                          "type": "integer"
                        },
                        "documents": {
                          "type": "integer"
                        }
                      }
                    },
                    "last_30_days": {
                      "type": "object",
                      "properties": {
                        "new_users": {
                          "type": "integer"
                        },
                        "new_documents": {
                          "type": "integer"
                        }
                      }
                    },
                    "role_distribution": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/admin/logs": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get audit logs (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "user_id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "action": {
                        "type": "string"
                      },
                      "timestamp": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "details": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/admin/logs/search": {
      "get": {
        "tags": ["Admin"],
        "summary": "Search audit logs (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "user_id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "action": {
                        "type": "string"
                      },
                      "timestamp": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "details": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    }
  }
} 